Add files via upload
Fixed errors. User is now able to set dates in Table A 2nd view controller which results in appearance of table view with exchange rates for chosen date range. 
2 buttons were added: close and refresh (when refresh button is clicked, data is being refreshed and spinner vc appears for a few seconds).
TO DO: Add the same features to Table B and C. Then work on UI to make it look better. Later - make the code shorter as well as more universal.
In the future, add some more features.